Fix applying too old backup for overlays (bsc#1070726)
In general problem is that system role selection clears overlay before
creating dialog, but it does it when going back and also when going
forward. So result is that when going back it is properly cleared and
then when going forward when different extensions is chosen, it is again
cleared, but with invalid backup. So this change in the end apply backup
only when going back.