Add a handler for licenses fetched from an RPM package
Basically, because it is necessary to know if the license acceptance
is required. The way to check it is looking for the
`no-acceptance-needed` file inside the RPM, which presence means that
the license does not need to be accepted.

This change also implies a few adjustments in the existing codebase.