Include the default lang in LicensesFetchers::Rpm#locales
Since it is possible to have the "LICENSE.TXT" file, without the
language code, which usually represents the license for the default
language.

This change is needed to keep consistent between the displayed license
content and the selected language in the license agreement dialog[1],
avoiding to select the first one because of the absence of proper value.

[1] https://github.com/yast/yast-packager/blob/e625f6ea375b9ffbc82eddae8774d066426b2f44/src/lib/y2packager/dialogs/inst_product_license.rb#L56