Fixed libzypp initialization (related to bsc#1199840)
In the YaST container read the configured repositories from
the host

- 4.5.8

Details:

It turned out that the problem is in the Pkg.GetSolverFlags call which is
called before initializing libzypp. In that case libzypp automatically
initializes itself, but it uses by default the system root (/) which is the
container. The libzypp initialization later done from YaST does not change
anything as libzypp is already initialized.

Also it initialized the libzypp in a wrong order. The libzypp target always
needs to be initialized before initializing the libzypp sources. The reason is
that the target defines which GPG keys and config file (zypp.conf) are used.
When initializing in wrong order you might get GPG signature errors.